putting trial_idx computation outside of feature_extraction
